requirements.txt
phonenumberslite
django-ckeditor
django-widget-tweaks
b2sdk
django-solo
booktype
imageio
pytils
zipstream
rawpy
redis
phonenumbers
chromedriver_autoinstaller
asgiref==3.9.1
attrs==25.3.0
boto3==1.40.31
botocore==1.40.31
brevo-python==1.2.0
browserforge==1.2.3
camoufox==0.4.11
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromedriver-autoinstaller==0.6.4
click==8.2.1
Django==5.2.5
django-allauth==65.10.0
django-background-tasks==1.2.8
django-s3direct==2.0.3
django-storages==1.14.6
djangorestframework==3.16.1
dropbox==12.0.2
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.10
jmespath==1.0.1
language-tags==1.2.0
lxml==6.0.1
numpy==2.3.2
orjson==3.11.3
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
ply==3.11
psycopg2-binary==2.9.10
pycparser==2.22
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.14.0
screeninfo==0.8.1
selenium==4.35.0
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
stone==3.3.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
undetected-chromedriver==3.5.5
unicorn==2.1.4
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
openai
pypdf
pdf2image